fix: made sidebar show the profile of the logged in user
