+++ src/widgets/Sidebar/model/types/sidebar.ts
@@ -1,0 +1,6 @@
+export interface SidebarItemType {
+  path: string;
+  text: string;
+  Icon: React.VFC<React.SVGProps<SVGSVGElement>>;
+  authOnly?: boolean;
+}
